Fix reading of TIFF files on 64 bit linux
* TIFFGetField TIFFTAG_STRIPBYTECOUNTS returns uint32**
  but plStripSize was of type long* which caused problems
  when compiling on 64 bit linux.

+++ datalib/nraDataCubeTIFFReader.h
@@ -190,7 +190,7 @@ public:
 	{
 	    
 	    long lStripNmbr;
-	    long* plStripSize;
+	    uint32_t* plStripSize;
 	    long lCurrStripSize;
 	    long lCurrRowsPerStrip;
 	    long lNumOfStrips;
